added get weather by location

+++ lib/models/weather_forecast_daily.dart
@@ -86,8 +86,8 @@ class Coord {
   Coord({required this.lon, required this.lat});
 
   Coord.fromJson(Map<String, dynamic> json) {
-    lon = json['lon'];
-    lat = json['lat'];
+    lon = json['lon'].toDouble();
+    lat = json['lat'].toDouble();
   }
 
   Map<String, dynamic> toJson() {
